--- a/webui/tests/cypress/integration/01-setup.spec.ts
+++ b/webui/tests/cypress/integration/01-setup.spec.ts
@@ -1,4 +1,6 @@
 describe('setup', () => {
+  const recordSelector = '#experimentsList tr.record';
+
   before(() => {
     cy.login();
   });
@@ -8,11 +10,11 @@ describe('setup', () => {
   });
 
   it('should have 4 experiments listed', () => {
-    cy.get('#experimentsList tr').should('have.lengthOf', 4);
+    cy.get(recordSelector).should('have.lengthOf', 4);
   });
 
   it('should have 4 active experiments listed', () => {
-    cy.get('#experimentsList tr').should('have.lengthOf', 4)
+    cy.get(recordSelector).should('have.lengthOf', 4)
       .each(($tr) => {
         cy.wrap($tr).should('contain', 'Active');
       });
@@ -24,7 +26,7 @@ describe('setup', () => {
     cy.get('.modal button').contains(/pause/i).click();
     /* eslint-disable-next-line cypress/no-unnecessary-waiting */
     cy.wait(5000);
-    cy.get('#experimentsList tr')
+    cy.get(recordSelector)
       .each(($tr) => {
         cy.wrap($tr).should('contain', 'Paused');
       });
@@ -32,26 +34,26 @@ describe('setup', () => {
   });
 
   it('should be able to unpause experiment 1', () => {
-    cy.get('#experimentsList tr td:nth-child(2)').contains('1').click();
+    cy.get(recordSelector + ' td:nth-child(2)').contains('1').click();
     cy.contains('Activate').click();
     cy.visit('/ui/experiments');
-    cy.get('#experimentsList tr').should('contain', 'Active');
+    cy.get(recordSelector).should('contain', 'Active');
   });
 
   it('should cancel experiment 2', () => {
-    cy.get('#experimentsList tr td:nth-child(2)').contains('2').click();
+    cy.get(recordSelector + ' td:nth-child(2)').contains('2').click();
     cy.contains('Cancel').click();
     cy.get('.modal button').contains(/cancel/i).click();
     cy.contains('Canceled', { timeout: Cypress.config('responseTimeout') });
   });
 
   it('should archive experiment 2', () => {
-    cy.get('#experimentsList tr td:nth-child(2)').contains('2').click();
+    cy.get(recordSelector + ' td:nth-child(2)').contains('2').click();
     cy.contains('Canceled');
     cy.get('body').should('not.contain', /archived/i);
     cy.contains('Archive').click();
     cy.contains(/archived/i);
     cy.visit('/ui/experiments');
-    cy.get('#experimentsList tr').should('have.lengthOf', 3);
+    cy.get(recordSelector).should('have.lengthOf', 3);
   });
 });