@hamidzr 2020-06-12
got an infrastructure failed error on CI https://app.circleci.com/pipelines/github/determined-ai/determined/2409/workflows/a714515a-a53b-49ab-8cba-1956218184e9/jobs/49115